refactor: Reorganiza pacotes de teste para seguir convenções do projeto
